Refactor: extract shared data logic to dedicated files
