+++ src/shared_data.h
@@ -1,0 +1,86 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <mutex>
+
+#include "include/proxy-wasm/context.h"
+#include "include/proxy-wasm/wasm.h"
+
+namespace proxy_wasm {
+
+class SharedData {
+public:
+  WasmResult get(std::string_view vm_id, const std::string_view key,
+                 std::pair<std::string, uint32_t> *result);
+
+  WasmResult set(std::string_view vm_id, std::string_view key, std::string_view value,
+                 uint32_t cas);
+
+  uint32_t registerQueue(std::string_view vm_id, std::string_view queue_name, uint32_t context_id,
+                         CallOnThreadFunction call_on_thread, std::string_view vm_key);
+  uint32_t resolveQueue(std::string_view vm_id, std::string_view queue_name);
+  WasmResult dequeue(uint32_t token, std::string *data);
+  WasmResult enqueue(uint32_t token, std::string_view value);
+
+  uint32_t nextCas() {
+    auto result = cas_;
+    cas_++;
+    if (!cas_) { // 0 is not a valid CAS value.
+      cas_++;
+    }
+    return result;
+  }
+
+private:
+  uint32_t nextQueueToken() {
+    while (true) {
+      uint32_t token = next_queue_token_++;
+      if (token == 0) {
+        continue; // 0 is an illegal token.
+      }
+      if (queue_token_set_.find(token) == queue_token_set_.end()) {
+        return token;
+      }
+    }
+  }
+
+  struct Queue {
+    std::string vm_key;
+    uint32_t context_id;
+    CallOnThreadFunction call_on_thread;
+    std::deque<std::string> queue;
+  };
+
+  // TODO: use std::shared_mutex in C++17.
+  std::mutex mutex_;
+  uint32_t cas_ = 1;
+  uint32_t next_queue_token_ = 1;
+  std::map<std::string, std::unordered_map<std::string, std::pair<std::string, uint32_t>>> data_;
+  std::map<uint32_t, Queue> queues_;
+  struct pair_hash {
+    template <class T1, class T2> std::size_t operator()(const std::pair<T1, T2> &pair) const {
+      return std::hash<T1>()(pair.first) ^ std::hash<T2>()(pair.second);
+    }
+  };
+  std::unordered_map<std::pair<std::string, std::string>, uint32_t, pair_hash> queue_tokens_;
+  std::unordered_set<uint32_t> queue_token_set_;
+};
+
+namespace {
+SharedData global_shared_data;
+} // namespace
+
+} // namespace proxy_wasm
